# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type MarketType = typeof MarketType[keyof typeof MarketType];


export const MarketType = {
  spot: 'spot',
  futures: 'futures',
} as const;

export interface HealthStatus {
  status: string;
}

export type SymbolParam = string;

export type IntervalParam = string;

/**
 * @minimum 1
 * @maximum 2000
 */
export type DaysParam = number;

export interface SymbolInfo {
  symbol: string;
  baseAsset: string;
  quoteAsset: string;
  status: string;
  minPrice: string;
  maxPrice: string;
  tickSize: string;
  minQty: string;
  stepSize: string;
}

export interface SymbolListResponse {
  symbols: SymbolInfo[];
  count: number;
}

export interface KlineCandle {
  openTime: number;
  closeTime: number;
  open: string;
  high: string;
  low: string;
  close: string;
  volume: string;
  quoteVolume: string;
  trades: number;
  takerBuyBaseVolume: string;
  takerBuyQuoteVolume: string;
}

export interface KlineResponse {
  symbol: string;
  interval: string;
  candles: KlineCandle[];
}

export interface TickerInfo {
  symbol: string;
  priceChange: string;
  priceChangePercent: string;
  weightedAvgPrice: string;
  prevClosePrice: string;
  lastPrice: string;
  lastQty: string;
  bidPrice: string;
  bidQty: string;
  askPrice: string;
  askQty: string;
  openPrice: string;
  highPrice: string;
  lowPrice: string;
  volume: string;
  quoteVolume: string;
  openTime: number;
  closeTime: number;
  firstId: number;
  lastId: number;
  count: number;
}

export interface TickerListResponse {
  tickers: TickerInfo[];
  count: number;
}

export type VwapLineValuesItem = {
  time: number;
  /** @nullable */
  value: number | null;
};

export interface VwapLine {
  name: string;
  color: string;
  values: VwapLineValuesItem[];
}

export type VwapBandUpperItem = {
  time: number;
  /** @nullable */
  value: number | null;
};

export type VwapBandLowerItem = {
  time: number;
  /** @nullable */
  value: number | null;
};

export interface VwapBand {
  name: string;
  upperColor: string;
  lowerColor: string;
  upper: VwapBandUpperItem[];
  lower: VwapBandLowerItem[];
}

export interface VwapSession {
  name: string;
  vwap: VwapLine;
  bands: VwapBand[];
}

export type VwapResponseDailyVwap = {
  current: VwapLine;
  previous: VwapLine;
  bands: VwapBand[];
};

export type VwapResponseWeeklyVwap = {
  current: VwapLine;
  previous: VwapLine;
  bands: VwapBand[];
};

export type VwapResponseDollarVolume = {
  perCandle: VwapLine;
  sma30: VwapLine;
  minimumThreshold: VwapLine;
  optimalThreshold: VwapLine;
};

export type VwapResponseSessionVolumeAccumulated = {
  accumulated: VwapLine;
  minimumThreshold: VwapLine;
  optimalThreshold: VwapLine;
};

export type VwapResponseRelativeQv = {
  relative: VwapLine;
  minimumThreshold: VwapLine;
};

export interface VwapResponse {
  symbol: string;
  interval: string;
  multiPeriodVwaps: VwapLine[];
  dailyVwap: VwapResponseDailyVwap;
  weeklyVwap: VwapResponseWeeklyVwap;
  sessions: VwapSession[];
  dollarVolume: VwapResponseDollarVolume;
  sessionVolumeAccumulated: VwapResponseSessionVolumeAccumulated;
  relativeQv: VwapResponseRelativeQv;
  /** VWMA auto-selected periods for current timeframe */
  vwapUltra1?: VwapLine[];
  /** VWMA Multi-Timeframe Map from higher timeframes */
  vwmaMtfMap?: VwapLine[];
  /** Z-Score oscillator (periods 48 and 84) */
  zScore?: VwapLine[];
}

export type MarketTypeParamParameter = MarketType;

export type ListSymbolsParams = {
/**
 * Binance Global market segment
 */
market?: MarketTypeParamParameter;
quote?: string;
};

export type GetHistoryParams = {
/**
 * Binance Global market segment
 */
market?: MarketTypeParamParameter;
symbol: SymbolParam;
interval?: IntervalParam;
days?: DaysParam;
};

export type GetVwapParams = {
/**
 * Binance Global market segment
 */
market?: MarketTypeParamParameter;
symbol: SymbolParam;
interval?: IntervalParam;
days?: DaysParam;
};

export type Get24hrTickerParams = {
/**
 * Binance Global market segment
 */
market?: MarketTypeParamParameter;
symbol?: SymbolParam;
};